Add RedirectConverterTest, cleanup RedirectListener.
